Corregido a fondo con valgrind
- Todo salvo still reachable que no importa y los avisos de
que sale de un proceso por un SIGNAL, lo cual está bien...
- Para correr como hice yo:
../build$ cmake -DCMAKE_BUILD_TYPE=Debug .. && make
../build$ valgrind --track-origins=yes --dsymutil=yes --leak-check=full ./tp_concurrente

Eso último va a mostrar los still reachable y los signal, ignorar
- También corregido casi todo warning de CLion (salvo algún never used).

+++ source/procesos/ProcesoGeneradorComensales.cpp
@@ -21,7 +21,7 @@ int ProcesoGeneradorComensales::ejecutarMiTarea() {
 
     while (!sigint_handler.getGracefulQuit()){
         /* Tiempo que transcurre hasta que llegue el próximo grupo */
-        int tiempoEspera = Utils::generarRandom(4);
+        unsigned tiempoEspera = (unsigned) Utils::generarRandom(4);
         sleep(tiempoEspera);
         //dormir(tiempoEspera); /* TODO Es un "slepp" sin sleep asi que vale */
         /* Cantidad de comensales en el grupo que llegó */
